refactor: changed widget name
